backend/pyproject.toml
[tool.poetry]
name = "movies-reco-backend"
version = "0.1.0"
description = "Flask API for Movies Recommendation"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
python-dotenv = "^1.0.1"
requests = "^2.32.0"
flask-cors = "^6.0.1"
cachetools = "^6.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
black = "^24.3.0"
isort = "^5.13.2"
pytest = "^8.0.0"
pytest-cov = "^5.0.0"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
